Refactor output components into separate file
