+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/text-generation-node-properties-panel/outputs/index.ts
@@ -2,3 +2,4 @@ export * from "./types";
 export * from "./utils";
 export * from "./use-connected-outputs";
 export * from "./use-categoried-outputs";
+export * from "./components";
